A Merge download model into setup script

+++ tutorials/download_model.sh
@@ -9,7 +9,7 @@ then
 fi
 
 # Get the bert model for the question answering demo
-${PYTHON} -m transformers.onnx --model=bert-large-uncased-whole-word-masking-finetuned-squad --feature=question-answering question_answering
+[ -f question_answering/model.onnx ] || ${PYTHON} -m transformers.onnx --model=bert-large-uncased-whole-word-masking-finetuned-squad --feature=question-answering question_answering
 
 # Get the gpt2 model for the generation model
-curl -fsSL -o generation/gpt2-lm-head-10.onnx https://github.com/onnx/models/raw/main/text/machine_comprehension/gpt-2/model/gpt2-lm-head-10.onnx
+[ -f generation/gpt2-lm-head-10.onnx ] || curl -fsSL -o generation/gpt2-lm-head-10.onnx https://github.com/onnx/models/raw/main/text/machine_comprehension/gpt-2/model/gpt2-lm-head-10.onnx
